+++ src/Xmobar/Plugins/Monitors/Disk.hs
@@ -31,26 +31,27 @@ data DiskIOOpts = DiskIOOpts
   { totalIconPattern :: Maybe IconPattern
   , writeIconPattern :: Maybe IconPattern
   , readIconPattern :: Maybe IconPattern
+  , contiguous :: Bool
   }
 
-parseDiskIOOpts :: [String] -> IO DiskIOOpts
-parseDiskIOOpts argv =
-  case getOpt Permute options argv of
-    (o, _, []) -> return $ foldr id defaultOpts o
-    (_, _, errs) -> ioError . userError $ concat errs
- where defaultOpts = DiskIOOpts
-          { totalIconPattern = Nothing
-          , writeIconPattern = Nothing
-          , readIconPattern = Nothing
-          }
-       options =
-          [ Option "" ["total-icon-pattern"] (ReqArg (\x o ->
-             o { totalIconPattern = Just $ parseIconPattern x}) "") ""
-          , Option "" ["write-icon-pattern"] (ReqArg (\x o ->
-             o { writeIconPattern = Just $ parseIconPattern x}) "") ""
-          , Option "" ["read-icon-pattern"] (ReqArg (\x o ->
-             o { readIconPattern = Just $ parseIconPattern x}) "") ""
-          ]
+dioDefaultOpts :: DiskIOOpts
+dioDefaultOpts = DiskIOOpts
+   { totalIconPattern = Nothing
+   , writeIconPattern = Nothing
+   , readIconPattern = Nothing
+   , contiguous = False
+   }
+
+dioOptions :: [OptDescr (DiskIOOpts -> DiskIOOpts)]
+dioOptions =
+   [ Option "" ["total-icon-pattern"] (ReqArg (\x o ->
+      o { totalIconPattern = Just $ parseIconPattern x}) "") ""
+   , Option "" ["write-icon-pattern"] (ReqArg (\x o ->
+      o { writeIconPattern = Just $ parseIconPattern x}) "") ""
+   , Option "" ["read-icon-pattern"] (ReqArg (\x o ->
+      o { readIconPattern = Just $ parseIconPattern x}) "") ""
+   , Option "c" ["contiguous"] (NoArg (\o -> o {contiguous = True})) ""
+   ]
 
 diskIOConfig :: IO MConfig
 diskIOConfig = mkMConfig "" ["total", "read", "write"
@@ -66,23 +67,24 @@ diskIOConfig = mkMConfig "" ["total", "read", "write"
 data DiskUOpts = DiskUOpts
   { freeIconPattern :: Maybe IconPattern
   , usedIconPattern :: Maybe IconPattern
+  , contiguousU :: Bool
   }
 
-parseDiskUOpts :: [String] -> IO DiskUOpts
-parseDiskUOpts argv =
-  case getOpt Permute options argv of
-    (o, _, []) -> return $ foldr id defaultOpts o
-    (_, _, errs) -> ioError . userError $ concat errs
- where defaultOpts = DiskUOpts
-          { freeIconPattern = Nothing
-          , usedIconPattern = Nothing
-          }
-       options =
-          [ Option "" ["free-icon-pattern"] (ReqArg (\x o ->
-             o { freeIconPattern = Just $ parseIconPattern x}) "") ""
-          , Option "" ["used-icon-pattern"] (ReqArg (\x o ->
-             o { usedIconPattern = Just $ parseIconPattern x}) "") ""
-          ]
+duDefaultOpts :: DiskUOpts
+duDefaultOpts = DiskUOpts
+   { freeIconPattern = Nothing
+   , usedIconPattern = Nothing
+   , contiguousU = False
+   }
+
+duOptions :: [OptDescr (DiskUOpts -> DiskUOpts)]
+duOptions =
+   [ Option "" ["free-icon-pattern"] (ReqArg (\x o ->
+      o { freeIconPattern = Just $ parseIconPattern x}) "") ""
+   , Option "" ["used-icon-pattern"] (ReqArg (\x o ->
+      o { usedIconPattern = Just $ parseIconPattern x}) "") ""
+   , Option "c" ["contiguous"] (NoArg (\o -> o {contiguousU = True})) ""
+   ]
 
 diskUConfig :: IO MConfig
 diskUConfig = mkMConfig ""
@@ -147,25 +149,26 @@ mountedData dref devs = do
   return $ map (parseDev (zipWith diff dt' dt)) devs
   where diff (dev, xs) (_, ys) = (dev, zipWith (-) xs ys)
 
+
 parseDev :: [(DevName, [Float])] -> DevName -> (DevName, [Float])
 parseDev dat dev =
   case find ((==dev) . fst) dat of
     Nothing -> (dev, [0, 0, 0])
     Just (_, xs) ->
-      let r = xs !! 2
-          w = xs !! 6
+      let r = 4096 * xs !! 2
+          w = 4096 * xs !! 6
           t = r + w
           rSp = speed r (xs !! 3)
           wSp = speed w (xs !! 7)
           sp =  speed t (xs !! 3 + xs !! 7)
-          speed x d = if d == 0 then 0 else 500 * x / d
+          speed x d = if d == 0 then 0 else x / d
           dat' = if length xs > 6
                  then [sp, rSp, wSp, t, r, w]
                  else [0, 0, 0, 0, 0, 0]
       in (dev, dat')
 
 speedToStr :: Float -> String
-speedToStr = showWithUnits 2 1
+speedToStr = showWithUnits 2 1 . (/ 1024)
 
 sizeToStr :: Integer -> String
 sizeToStr = showWithUnits 3 0 . fromIntegral
@@ -201,11 +204,11 @@ runDiskIO' opts (tmp, xs) = do
 
 runDiskIO :: DevDataRef -> [(String, String)] -> [String] -> Monitor String
 runDiskIO dref disks argv = do
-  opts <- io $ parseDiskIOOpts argv
+  opts <- io $ parseOptsWith dioOptions dioDefaultOpts argv
   dev <- io $ mountedOrDiskDevices (map fst disks)
   dat <- io $ mountedData dref (map fst dev)
   strs <- mapM (runDiskIO' opts) $ devTemplates disks dev dat
-  return $ unwords strs
+  return $ (if contiguous opts then concat else unwords) strs
 
 startDiskIO :: [(String, String)] ->
                [String] -> Int -> (String -> IO ()) -> IO ()
@@ -247,6 +250,6 @@ runDiskU' opts tmp path = do
 runDiskU :: [(String, String)] -> [String] -> Monitor String
 runDiskU disks argv = do
   devs <- io $ mountedDevices (map fst disks)
-  opts <- io $ parseDiskUOpts argv
+  opts <- io $ parseOptsWith duOptions duDefaultOpts argv
   strs <- mapM (\(d, p) -> runDiskU' opts (findTempl d p disks) p) devs
-  return $ unwords strs
+  return $ (if contiguousU opts then concat else unwords) strs